requirements.txt
Flask>=3.0.0
numpy>=1.26.0
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
